+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./CodeFile";
 export * from "./CodeBuilder";
 export * from "./types/ManualSectionMap";
+export * from "./types/CommentSyntax";